+++ torch/csrc/jit/python_tracer.cpp
@@ -51,6 +51,9 @@ void initPythonTracerBindings(PyObject* module_) {
       for (auto& kv : export_map) {
         auto t = kv.second;
         size_t copy_bytes = t.type().elementSizeInBytes() * t.numel();
+        // TODO: this is an unecessary copy. In theory we can directly return
+        // the map from identifier to Tensor, but we need some API in Python
+        // to get raw `bytes` containing the raw tensor data.
         python_serialized_export_map[kv.first] = py::bytes(static_cast<const char*>(t.data_ptr()), copy_bytes);
       }
       return std::make_tuple(
